+++ src/app/components/App/index.js
@@ -7,15 +7,19 @@ import Settings from '../../containers/Settings'
 
 import './styles.css'
 
-const App = ({ coffeeScriptVersion, output, status }) => {
+const App = ({ coffeeScriptVersion, editorTheme, outputTheme, output, status }) => {
   return (
     <div className='App'>
       <div className='App__Settings'>
-        <Settings coffeeScriptVersion={coffeeScriptVersion} />
+        <Settings
+          coffeeScriptVersion={coffeeScriptVersion}
+          editorTheme={editorTheme}
+          outputTheme={outputTheme}
+        />
       </div>
       <div className='App__EditorOutput'>
-        <Editor coffeeScriptVersion={coffeeScriptVersion} />
-        <Output>{output}</Output>
+        <Editor coffeeScriptVersion={coffeeScriptVersion} theme={editorTheme} />
+        <Output theme={outputTheme}>{output}</Output>
       </div>
       <div className='App__Status'>
         <Status status={status} />
@@ -27,6 +31,8 @@ const App = ({ coffeeScriptVersion, output, status }) => {
 const mapStateToProps = (state) => {
   return {
     coffeeScriptVersion: state.app.coffeeScriptVersion,
+    editorTheme: state.app.editorTheme,
+    outputTheme: state.app.outputTheme,
     output: state.app.output,
     status: state.app.status
   }