+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 1.1.0
 
 - Set the version of CoffeeScript to target
+- Set the theme for the editor
+- Set the theme for the output
 
 # 1.0.0